Cassandane: don't let TestEntity JMAP use affect CreatedIds
Mail::JMAPTalk has an odd behavior where any creationIds provided by
call 1 will be passed back in to call 2.  This is implicit and stored on
the client, which means that "back office" use of the client can confuse
the state of things.  I'd prefer to avoid this behavior entirely, but
for now, this weird localization should solve the problem.