message.c: set numpart field to 1 for cached RFC822 bodies
This patch changes the message cache reading code to set the
body->numparts field to one for embedded RFC 822 messages.

Before, when reading a cached message with message_read_bodystructure,
the body->subpart field for a body part of type "message/rfc822"
was set but the body->numparts field was kept initialized with zero.
In contrast, if the same cached message was read using the now removed
message_parse_cbodystructure function, the field was set to one.

It's unclear if this was a deliberate choice and likely was not:

Keeping numparts set to zero required any code that traverses a struct
body tree to special-handle body parts of type "message/rfc822", as
descending the tree as long as body->numparts is non-zero did stop at
an embededded message. Also, the same body structure was represented
differently depending on if it was read with message_read_bodystructure
or message_parse_cbodystructure.