Cassandane test entities: load AutoSetup after defining methods
Before this commit code did this:

    package XYZ::Factory::ContactCard;
    use Cassandane::TestEntity::AutoSetup;
    sub fill_in_creation_defaults {...}

AutoSetup provided sugar for something like this:

    package XYZ::Factory::ContactCard;
    use Moo;
    with 'Cassandane::TestEntity::Role::Factory';
    # other stuff
    sub fill_in_creation_defaults {...}

In the unsugared scenario, `with` comes before `sub ...` in the text,
but occurred at runtime, while subroutine definitions are compiled at
compile time.  That meant that when `with` was evaluated, the role
applicator could determine that the composing class already had a
subroutine with that name and decline to install it.  (In-class
methods prevent role methods from being installed.)

In the sugared version, the `with` was applied by the implicit call to
AutoSetup->import, which occurs at compile time.  Then the
fill_in_creation_defaults method gets installed by role application and
is then *reinstalled* by the declaration in the class.  This issues a
"redefine"-class warning.  AutoSetup can't trivially defer role
application, so instead we move it to the end of the class.  Problem
solved.