WIP: Cassandane: structured vCard testing
This commit isn't quite done yet, but gets the point across.

First, it adds ->as_vcardX_struct methods to ContactCard instances,
which return single-vcard-object result of calling
Test::VCardFast::vcard2hash on a vcard blob.  Then one test is tweaked
to use the resulting structure as the basis for testing card content.

Probably we should improve the "value" subroutine that's used to
generate simple property value tests so that we can use it everywhere.
As is, it's okay, but isn't not quite exensible enough for general use.

If we like this general idea, I'll rework this more.