Cassandane: switch TestEntity properties to lazy
This copies more from Fastmail's Newt system.  I have mixed feelings,
but "actually gets the job done" wins out over "maximally efficient".

This turns the property hashref into a lazy attribute that can be
cleared and re-retrieved.  A factory-powered ->get will populate it with
the gotten properties, but an ->update will clear those so that, when
read, they will be re-retrieved.

The up side:  Updating an instance with a patch like
`{ "shareWith/identifier/mayVerb": true }` and then reading the
shareWith property will provide the correct value, because it will
re-read the property from the server.

The down side: `$instance->someProperty(123); $instance->someProperty`
will now result in two JMAP calls, because the first will clear the
property hash and the second will do a Thing/get to re-populate it.
This could've been prevented by allowing the mutator to patch the
properties directly.  We may re-add that optimization selectively in the
future.  For now, "correct" beats out "fast".