twom: add TWOM_ONELOCK flag and use it for some internal re-locking
This means that we don't take the headlock multiple times.  A process
which has already been through the headlock dance once for this
action SHOULD get higher priority, so allow it to compete directly
on the datalock.

This is safe, because ONLY the datalock protects the actual data,
and the headlock is just a gate mechanism to stop new processes from
coming along and reading while a writer is waiting.

This favours processes in the middle of a foreach somewhat, and
reduces the number of syscalls significantly