github-actions: enable and run Test::Core tests in CI
This arranges for the docker container to run in privileged mode,
which lets us set the kernel.core_pattern sysctl to a value that
Cassandane knows how to work with.

The Test::Core tests aren't directly interesting, but if they can
run and pass, that proves that Cassandane can detect when a core
file was generated by a running test.  Tests that generate core
files automatically fail.

This means PRs that introduce crash bugs will now fail in CI...
as long as the crashing code is exercised by a test, anyway.