carddav_db.c: find lowest PREF value and flag as default email
When writing a card out, we mistakenly recorded whatever emails(s) had a `pref` value of exactly `1` as the preferred emails, and later when fetching them (in, say, expandcard) we'd pick one of those as the default email.

What we're supposed to do is pick the lowest `pref` value (from 1-100) as the preferred email(s), and then select one
of those as the default email; so at card write time we will now do just that. If there are no explicitly preferred emails, we will pick one out of the entire list.

This solves a bug where if you create an email with a `pref` of >1, and that email is the preferred email, it wasn't
being treated as the default email since we were ignoring any `pref` value but `1`.