Cassandane::JMAPTester: add, and use, ->set_username_and_password
In tests that rename users, we want to update the JMAP client to use the
new username for authentication.  The old code went fiddling around in
the guts of Mail::JMAPTalk.  That will not stand!  Just move the auth
header setup into Cassandane::JMAPTester and use it from TestUser's
setup.