Cassandane: replace literal JSON with use of ->CallMethods
These tests constructed JSON by hand, but for no clear or documented
reason.  The behavior they test is about the structure of mailboxes
described by the JSON, not the ordering of elements in the JSON string.
Converting these to use normal JSON generation and ->CallMethods or
->http_request makes the code shorter and clearer.

We had some discussion about the change from embedded JSON to
Perl-generated JSON, because this was unusual.  Bron, the original
author of the "mailbox_move_to_deleted_parent" test, felt it would be
fine to generate the JSON from Perl.