FASTMAIL ONLY - http_carddav.c: allow changing v3 UID to v4 UID in PUT
This allows CardDAV PUTs that rewrite a version 3 vCard to version 4
and prefix the UID value of the former card with the verbatim string
"urn:uuid".

Strictly speaking, this is not allowed and would require us to reject
this with a no-uid-conflict precondition error. But previous versions
of Cyrus did allow to do so and there is at least one client (eMClient)
that does these bogus rewrites, so let's allow this case until they got
a chance to fix their code.