Cassandane: merge paired test entity classes single modules
Each type of test entity has a Factory class and an Instance class.
Previously, each one was in its own module.  This was perfectly fine and
sensible, but I'd like it to be easy for people to read the docs on how
to use test entities, and figuring out whether to look at one module or
the other seems tedious.  Instead, we put both classes in one module,
and we'll document *those*.